+++ changelog.md
@@ -1,5 +1,6 @@
 # ChangeLog
 
+## Version 5.2.11 (Aug 31st 2015)
 * Don't switch to quoted-printable for long lines if already using base64
 * Fixed Travis-CI config when run on PHP 7
 * Added Google XOAUTH2 authentication mechanism, thanks to @sherryl4george